LC-number-of-smooth-descent-periods-of-a-stock [Runtime: 7 ms; Memory: 107 MB]

+++ array/number-of-smooth-descent-periods-of-a-stock.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     long long getDescentPeriods(vector<int>& prices) {
-        int ans=prices.size();
+        long long ans=prices.size();
 
-        int k=1;
+        long long k=1;
 
         for(int i=0; i<prices.size()-1; i++){
             if(prices[i]-prices[i+1]==1){
